+++ app/src/main/java/com/github/HumanLearning2021/HumanLearningApp/room/RoomDataset.kt
@@ -1,7 +1,6 @@
 package com.github.HumanLearning2021.HumanLearningApp.room
 
-import androidx.room.Embedded
-import androidx.room.Relation
+import androidx.room.*
 
 data class RoomDataset(
     @Embedded val datasetWithoutCategories: RoomDatasetWithoutCategories,
@@ -10,3 +9,30 @@ data class RoomDataset(
         entityColumn = "categoryId",
     ) val category: RoomCategory?
 )
+
+@Dao
+interface DatasetDao {
+    @Transaction
+    @Query("SELECT * FROM dataset")
+    fun loadAll(): List<RoomDataset>
+
+    @Transaction
+    @Query("SELECT * FROM dataset WHERE datasetId = :id LIMIT 1")
+    fun loadById(id: String): RoomDataset
+
+    @Transaction
+    @Query("SELECT * FROM dataset WHERE name = :name")
+    fun loadByName(name: String): List<RoomDataset>
+
+    @Insert
+    fun insertAll(vararg datasetWithoutCategories: RoomDatasetWithoutCategories)
+
+    @Insert
+    fun insertAll(vararg categories: RoomCategory)
+
+    @Delete
+    fun delete(datasetWithoutCategories: RoomDatasetWithoutCategories)
+
+    @Delete
+    fun delete(category: RoomCategory)
+}
